[Transport] Refactor Consumer to share code
