+++ faust/transport/consumer.py
@@ -68,10 +68,13 @@ from typing import (
 )
 from weakref import WeakSet
 
-from mode import Service, ServiceT, get_logger
+from mode import Service, ServiceT, flight_recorder, get_logger
+from mode.threads import MethodQueue, QueueServiceThread
 from mode.utils.futures import notify
+from mode.utils.locks import Event
+from mode.utils.times import Seconds
 from faust.exceptions import ProducerSendError
-from faust.types import AppT, Message, TP
+from faust.types import AppT, ConsumerMessage, Message, TP
 from faust.types.transports import (
     ConsumerCallback,
     ConsumerT,
@@ -80,7 +83,10 @@ from faust.types.transports import (
     TPorTopicSet,
     TransportT,
 )
+from faust.utils import terminal
 from faust.utils.functional import consecutive_numbers
+
+from .utils import TopicBuffer, TopicIndexMap
 
 if typing.TYPE_CHECKING:  # pragma: no cover
     from faust.app import App
@@ -99,6 +105,16 @@ CONSUMER_SEEKING = 'SEEKING'
 CONSUMER_WAIT_EMPTY = 'WAIT_EMPTY'
 
 logger = get_logger(__name__)
+
+RecordMap = Mapping[TP, List[Any]]
+
+
+def ensure_TP(tp: Any) -> TP:
+    return tp if isinstance(tp, TP) else TP(tp.topic, tp.partition)
+
+
+def ensure_TPset(tps: Iterable[Any]) -> Set[TP]:
+    return {ensure_TP(tp) for tp in tps}
 
 
 class Fetcher(Service):
@@ -194,6 +210,12 @@ class Consumer(Service, ConsumerT):
     _commit_every: Optional[int]
     _n_acked: int = 0
 
+    _active_partitions: Optional[Set[TP]]
+    _paused_partitions: Set[TP]
+
+    flow_active: bool = True
+    can_resume_flow: Event
+
     def __init__(self,
                  transport: TransportT,
                  callback: ConsumerCallback,
@@ -227,13 +249,93 @@ class Consumer(Service, ConsumerT):
         self._last_batch = None
         self._end_offset_monitor_interval = self.commit_interval * 2
         self.randomly_assigned_topics = set()
+        self.can_resume_flow = Event()
+        self._reset_state()
         super().__init__(loop=loop or self.transport.loop, **kwargs)
+
+    def _reset_state(self) -> None:
+        self._active_partitions = None
+        self._paused_partitions = set()
+        self.can_resume_flow.clear()
+        self.flow_active = True
+        self._last_batch = None
+        self._time_start = monotonic()
+
+    async def on_restart(self) -> None:
+        self._reset_state()
+        self.on_init()
+
+    def _get_active_partitions(self) -> Set[TP]:
+        tps = self._active_partitions
+        if tps is None:
+            return self._set_active_tps(self.assignment())
+        assert all(isinstance(x, TP) for x in tps)
+        return tps
+
+    def _set_active_tps(self, tps: Set[TP]) -> Set[TP]:
+        xtps = self._active_partitions = ensure_TPset(tps)  # copy
+        xtps.difference_update(self._paused_partitions)
+        return xtps
+
+    def _records_to_topic_index(
+            self,
+            records: RecordMap,
+            active_partitions: Set[TP]) -> TopicIndexMap:
+        return TopicBuffer.map_from_records(records)
 
     @abc.abstractmethod
     async def _commit(
             self,
-            offsets: Mapping[TP, Tuple[int, str]]) -> bool:  # pragma: no cover
+            offsets: Mapping[TP, int]) -> bool:  # pragma: no cover
         ...
+
+    async def perform_seek(self) -> None:
+        read_offset = self._read_offset
+        _committed_offsets = await self.seek_to_committed()
+        committed_offsets = {
+            ensure_TP(tp): offset
+            for tp, offset in _committed_offsets.items()
+            if offset is not None
+        }
+        read_offset.update({
+            tp: offset if offset else None
+            for tp, offset in committed_offsets.items()
+        })
+        self._committed_offset.update(committed_offsets)
+
+    @abc.abstractmethod
+    async def seek_to_committed(self) -> Mapping[TP, int]:
+        ...
+
+    async def seek(self, partition: TP, offset: int) -> None:
+        self.log.dev('SEEK %r -> %r', partition, offset)
+        # reset livelock detection
+        self._last_batch = None
+        # set new read offset so we will reread messages
+        self._read_offset[ensure_TP(partition)] = offset if offset else None
+        self._seek(partition, offset)
+
+    @abc.abstractmethod
+    async def _seek(self, partition: TP, offset: int) -> None:
+        ...
+
+    def stop_flow(self) -> None:
+        self.flow_active = False
+        self.can_resume_flow.clear()
+
+    def resume_flow(self) -> None:
+        self.flow_active = True
+        self.can_resume_flow.set()
+
+    def pause_partitions(self, tps: Iterable[TP]) -> None:
+        tpset = ensure_TPset(tps)
+        self._get_active_partitions().difference_update(tpset)
+        self._paused_partitions.update(tpset)
+
+    def resume_partitions(self, tps: Iterable[TP]) -> None:
+        tpset = ensure_TPset(tps)
+        self._get_active_partitions().update(tps)
+        self._paused_partitions.difference_update(tpset)
 
     @abc.abstractmethod
     def _new_topicpartition(
@@ -243,13 +345,131 @@ class Consumer(Service, ConsumerT):
     def _is_changelog_tp(self, tp: TP) -> bool:
         return tp.topic in self.app.tables.changelog_topics
 
-    @Service.transitions_to(CONSUMER_PARTITIONS_ASSIGNED)
-    async def on_partitions_assigned(self, assigned: Set[TP]) -> None:
-        await self._on_partitions_assigned(assigned)
-
     @Service.transitions_to(CONSUMER_PARTITIONS_REVOKED)
     async def on_partitions_revoked(self, revoked: Set[TP]) -> None:
+        self.app.rebalancing = True  # set as early as possible
+        # see comment in on_partitions_assigned
+        # remove revoked partitions from active + paused tps.
+        if self._active_partitions is not None:
+            self._active_partitions.difference_update(revoked)
+        self._paused_partitions.difference_update(revoked)
         await self._on_partitions_revoked(revoked)
+
+    @Service.transitions_to(CONSUMER_PARTITIONS_ASSIGNED)
+    async def on_partitions_assigned(self, assigned: Set[TP]) -> None:
+        # remove recently revoked tps from set of paused tps.
+        self._paused_partitions.intersection_update(assigned)
+        # cache set of assigned partitions
+        self._set_active_tps(assigned)
+        # start callback chain of assigned callbacks.
+        #   need to copy set at this point, since we cannot have
+        #   the callbacks mutate our active list.
+        self._last_batch = None
+        await self._on_partitions_assigned(assigned)
+
+    @abc.abstractmethod
+    async def _getmany(self,
+                       active_partitions: Set[TP],
+                       timeout: float) -> RecordMap:
+        ...
+
+    async def getmany(self,
+                      timeout: float) -> AsyncIterator[Tuple[TP, Message]]:
+        if not self.flow_active:
+            await self.wait(self.can_resume_flow)
+        # Implementation for the Fetcher service.
+        active_partitions = self._get_active_partitions()
+        _next = next
+
+        records: RecordMap = {}
+        if active_partitions:
+            # Fetch records only if active partitions to avoid the risk of
+            # fetching all partitions in the beginning when none of the
+            # partitions is paused/resumed.
+            records = await self._getmany(
+                active_partitions=active_partitions,
+                timeout=timeout,
+            )
+        else:
+            # We should still release to the event loop
+            await self.sleep(1)
+            if self.should_stop:
+                return
+
+        # records' contain mapping from TP to list of messages.
+        # if there are two agents, consuming from topics t1 and t2,
+        # normal order of iteration would be to process each
+        # tp in the dict:
+        #    for tp. messages in records.items():
+        #        for message in messages:
+        #           yield tp, message
+        #
+        # The problem with this, is if we have prefetched 16k records
+        # for one partition, the other partitions won't even start processing
+        # before those 16k records are completed.
+        #
+        # So we try round-robin between the tps instead:
+        #
+        #    iterators: Dict[TP, Iterator] = {
+        #        tp: iter(messages)
+        #        for tp, messages in records.items()
+        #    }
+        #    while iterators:
+        #        for tp, messages in iterators.items():
+        #            yield tp, next(messages)
+        #            # remove from iterators if empty.
+        #
+        # The problem with this implementation is that
+        # the records mapping is ordered by TP, so records.keys()
+        # will look like this:
+        #
+        #  TP(topic='bar', partition=0)
+        #  TP(topic='bar', partition=1)
+        #  TP(topic='bar', partition=2)
+        #  TP(topic='bar', partition=3)
+        #  TP(topic='foo', partition=0)
+        #  TP(topic='foo', partition=1)
+        #  TP(topic='foo', partition=2)
+        #  TP(topic='foo', partition=3)
+        #
+        # If there are 100 partitions for each topic,
+        # it will process 100 items in the first topic, then 100 items
+        # in the other topic, but even worse if partition counts
+        # vary greatly, t1 has 1000 partitions and t2
+        # has 1 partition, then t2 will end up being starved most of the time.
+        #
+        # We solve this by going round-robin through each topic.
+        topic_index = self._records_to_topic_index(records, active_partitions)
+        to_remove: Set[str] = set()
+        sentinel = object()
+
+        to_message = self._to_message  # localize
+
+        while topic_index:
+            if not self.flow_active:
+                break
+            for topic in to_remove:
+                topic_index.pop(topic, None)
+            for topic, messages in topic_index.items():
+                if not self.flow_active:
+                    break
+                item = _next(messages, sentinel)
+                if item is sentinel:
+                    # this topic is now empty,
+                    # but we cannot remove from dict while iterating over it,
+                    # so move that to the outer loop.
+                    to_remove.add(topic)
+                    continue
+                tp, record = item  # type: ignore
+                if tp in active_partitions:
+                    highwater_mark = self.highwater(tp)
+                    self.app.monitor.track_tp_end_offset(tp, highwater_mark)
+                    # convert timestamp to seconds from int milliseconds.
+                    yield tp, to_message(tp, record)
+
+    @abc.abstractmethod
+    def _to_message(self, tp: TP, record: Any) -> ConsumerMessage:
+        ...
 
     def track_message(self, message: Message) -> None:
         # add to set of pending messages that must be acked for graceful
@@ -316,6 +536,8 @@ class Consumer(Service, ConsumerT):
     async def on_stop(self) -> None:
         if self.app.conf.stream_wait_empty:
             await self.wait_empty()
+        else:
+            await self.commit()
         self._last_batch = None
 
     @Service.task
@@ -427,12 +649,38 @@ class Consumer(Service, ConsumerT):
             if pending:
                 await producer.wait_many(pending)
 
-    async def _commit_offsets(self, commit_offsets: Mapping[TP, int]) -> bool:
-        meta = ''
-        return await self._commit({
-            tp: (offset, meta)
-            for tp, offset in commit_offsets.items()
-        })
+    async def _commit_offsets(self, offsets: Mapping[TP, int]) -> bool:
+        table = terminal.logtable(
+            [(str(tp), str(offset))
+             for tp, offset in offsets.items()],
+            title='Commit Offsets',
+            headers=['TP', 'Offset'],
+        )
+        self.log.dev('COMMITTING OFFSETS:\n%s', table)
+        assignment = self.assignment()
+        commitable_offsets: Dict[TP, int] = {}
+        revoked: Dict[TP, int] = {}
+        for tp, offset in offsets.items():
+            if tp in assignment:
+                commitable_offsets[tp] = offset
+            else:
+                revoked[tp] = offset
+        if revoked:
+            self.log.info(
+                'Discarded commit for revoked partitions that '
+                'will be eventually processed again: %r',
+                revoked,
+            )
+        if not commitable_offsets:
+            return False
+        with flight_recorder(self.log, timeout=300.0) as on_timeout:
+            on_timeout.info('+consumer.commit()')
+            await self._commit(commitable_offsets)
+            on_timeout.info('-consumer.commit()')
+        self._committed_offset.update(commitable_offsets)
+        self.app.monitor.on_tp_commit(commitable_offsets)
+        self._last_batch = None
+        return True
 
     def _filter_tps_with_pending_acks(
             self, topics: TPorTopicSet = None) -> Iterator[TP]:
@@ -537,3 +785,184 @@ class Consumer(Service, ConsumerT):
     @property
     def unacked(self) -> Set[Message]:
         return cast(Set[Message], self._unacked_messages)
+
+
+class ConsumerThread(QueueServiceThread):
+    app: AppT
+    consumer: 'ThreadDelegateConsumer'
+    transport: TransportT
+
+    def __init__(self, consumer: ConsumerT, **kwargs: Any) -> None:
+        self.consumer = consumer
+        self.transport = self.consumer.transport
+        self.app = self.transport.app
+        super().__init__(**kwargs)
+
+    @abc.abstractmethod
+    async def subscribe(self, topics: Iterable[str]) -> None:
+        ...
+
+    @abc.abstractmethod
+    async def seek_to_committed(self) -> Mapping[TP, int]:
+        ...
+
+    @abc.abstractmethod
+    async def commit(self, tps: Mapping[TP, int]) -> bool:
+        ...
+
+    @abc.abstractmethod
+    async def position(self, tp: TP) -> Optional[int]:
+        ...
+
+    @abc.abstractmethod
+    async def seek_to_beginning(self, *partitions: TP) -> None:
+        ...
+
+    @abc.abstractmethod
+    async def seek_wait(self, partitions: Mapping[TP, int]) -> None:
+        ...
+
+    @abc.abstractmethod
+    def seek(self, partition: TP, offset: int) -> None:
+        ...
+
+    @abc.abstractmethod
+    def assignment(self) -> Set[TP]:
+        ...
+
+    @abc.abstractmethod
+    def highwater(self, tp: TP) -> int:
+        ...
+
+    @abc.abstractmethod
+    def topic_partitions(self, topic: str) -> Optional[int]:
+        ...
+
+    @abc.abstractmethod
+    async def earliest_offsets(self,
+                               *partitions: TP) -> MutableMapping[TP, int]:
+        ...
+
+    @abc.abstractmethod
+    async def highwaters(self, *partitions: TP) -> MutableMapping[TP, int]:
+        ...
+
+    @abc.abstractmethod
+    async def getmany(self,
+                      active_partitions: Set[TP],
+                      timeout: float) -> RecordMap:
+        ...
+
+    @abc.abstractmethod
+    async def create_topic(self,
+                           topic: str,
+                           partitions: int,
+                           replication: int,
+                           *,
+                           config: Mapping[str, Any] = None,
+                           timeout: Seconds = 30.0,
+                           retention: Seconds = None,
+                           compacting: bool = None,
+                           deleting: bool = None,
+                           ensure_created: bool = False) -> None:
+        ...
+
+    async def on_partitions_revoked(
+            self, revoked: Set[TP]) -> None:
+        await self.consumer.threadsafe_partitions_revoked(
+            self.thread_loop, revoked)
+
+    async def on_partitions_assigned(
+            self, assigned: Set[TP]) -> None:
+        await self.consumer.threadsafe_partitions_assigned(
+            self.thread_loop, assigned)
+
+
+class ThreadDelegateConsumer(Consumer):
+
+    _thread: ConsumerThread
+
+    #: Main thread method queue.
+    #: The consumer is running in a separate thread, and so we send
+    #: requests to it via a queue.
+    #: Sometimes the thread needs to call code owned by the main thread,
+    #: such as App.on_partitions_revoked, and in that case the thread
+    #: uses this method queue.
+    _method_queue: MethodQueue
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super().__init__(*args, **kwargs)
+        self._method_queue = MethodQueue(loop=self.loop, beacon=self.beacon)
+        self.add_dependency(self._method_queue)
+        self._thread = self._new_consumer_thread()
+        self.add_dependency(self._thread)
+
+    @abc.abstractmethod
+    def _new_consumer_thread(self) -> ConsumerThread:
+        ...
+
+    async def threadsafe_partitions_revoked(
+            self,
+            receiver_loop: asyncio.AbstractEventLoop,
+            revoked: Set[TP]) -> None:
+        promise = await self._method_queue.call(
+            receiver_loop.create_future(),
+            self.on_partitions_revoked,
+            revoked,
+        )
+        # wait for main-thread to finish processing request
+        await promise
+
+    async def threadsafe_partitions_assigned(
+            self,
+            receiver_loop: asyncio.AbstractEventLoop,
+            assigned: Set[TP]) -> None:
+        promise = await self._method_queue.call(
+            receiver_loop.create_future(),
+            self.on_partitions_assigned,
+            assigned,
+        )
+        # wait for main-thread to finish processing request
+        await promise
+
+    async def _getmany(self,
+                       active_partitions: Set[TP],
+                       timeout: float) -> RecordMap:
+        return await self._thread.getmany(active_partitions, timeout)
+
+    async def subscribe(self, topics: Iterable[str]) -> None:
+        await self._thread.subscribe(topics=topics)
+
+    async def seek_to_committed(self) -> Mapping[TP, int]:
+        return await self._thread.seek_to_committed()
+
+    async def position(self, tp: TP) -> Optional[int]:
+        return await self._thread.position(tp)
+
+    async def seek_wait(self, partitions: Mapping[TP, int]) -> None:
+        return await self._thread.seek_wait(partitions)
+
+    async def _seek(self, partition: TP, offset: int) -> None:
+        self._thread.seek(partition, offset)
+
+    def assignment(self) -> Set[TP]:
+        return self._thread.assignment()
+
+    def highwater(self, tp: TP) -> int:
+        return self._thread.highwater(tp)
+
+    def topic_partitions(self, topic: str) -> Optional[int]:
+        return self._thread.topic_partitions(topic)
+
+    async def earliest_offsets(self,
+                               *partitions: TP) -> MutableMapping[TP, int]:
+        return await self._thread.earliest_offsets(*partitions)
+
+    async def highwaters(self, *partitions: TP) -> MutableMapping[TP, int]:
+        return await self._thread.highwaters(*partitions)
+
+    async def _commit(self, offsets: Mapping[TP, int]) -> bool:
+        return await self._thread.commit(offsets)
+
+    def close(self) -> None:
+        self._thread.close()
