+++ faust/transport/drivers/aiokafka.py
@@ -2,12 +2,9 @@
 import asyncio
 from typing import (
     Any,
-    AsyncIterator,
     Awaitable,
     ClassVar,
-    Dict,
     Iterable,
-    Iterator,
     List,
     Mapping,
     MutableMapping,
@@ -15,7 +12,6 @@ from typing import (
     Set,
     Tuple,
     Type,
-    Union,
     cast,
 )
 
@@ -28,7 +24,6 @@ from aiokafka.errors import (
     KafkaError,
 )
 from aiokafka.structs import (
-    ConsumerRecord,
     OffsetAndMetadata,
     TopicPartition as _TopicPartition,
 )
@@ -39,19 +34,21 @@ from rhkafka.errors import (
 )
 from rhkafka.partitioner.default import DefaultPartitioner
 from rhkafka.protocol.metadata import MetadataRequest_v1
-from mode import Service, flight_recorder, get_logger
-from mode.threads import MethodQueue, QueueServiceThread
-from mode.utils.compat import OrderedDict
+from mode import Service, get_logger
 from mode.utils.futures import StampedeWrapper
-from mode.utils.locks import Event
 from mode.utils.times import Seconds, want_seconds
 from yarl import URL
 
 from faust.exceptions import ConsumerNotStarted, ProducerSendError
 from faust.transport import base
-from faust.types import AppT, ConsumerMessage, Message, RecordMetadata, TP
+from faust.transport.consumer import (
+    ConsumerThread,
+    RecordMap,
+    ThreadDelegateConsumer,
+    ensure_TPset,
+)
+from faust.types import ConsumerMessage, RecordMetadata, TP
 from faust.types.transports import ConsumerT, ProducerT
-from faust.utils import terminal
 from faust.utils.kafka.protocol.admin import CreateTopicsRequest
 
 __all__ = ['Consumer', 'Producer', 'Transport']
@@ -59,19 +56,6 @@ __all__ = ['Consumer', 'Producer', 'Transport']
 if not hasattr(aiokafka, '__robinhood__'):
     raise RuntimeError(
         'Please install robinhood-aiokafka, not aiokafka')
-
-# This is what we get from aiokafka getmany()
-# A mapping of TP to buffer-list of records.
-RecordMap = Mapping[TP, List[ConsumerRecord]]
-
-# But we want to process records from topics in round-robin order.
-# We convert records into a mapping from topic-name to "chain-of-buffers":
-#   topic_index['topic-name'] = chain(all_topic_partition_buffers)
-# This means we can get the next message available in any topic
-# by doing: next(topic_index['topic_name'])
-TopicIndexMap = MutableMapping[str, '_TopicBuffer']
-
-_TPTypes = Union[TP, _TopicPartition]
 
 logger = get_logger(__name__)
 
@@ -81,120 +65,95 @@ def server_list(urls: List[URL], default_port: int) -> List[str]:
     return [f'{u.host or default_host}:{u.port or default_port}' for u in urls]
 
 
-def _ensure_TP(tp: _TPTypes) -> TP:
-    return tp if isinstance(tp, TP) else TP(tp.topic, tp.partition)
-
-
-class _TopicBuffer(Iterator):
-    _buffers: Dict[TP, Iterator[ConsumerRecord]]
-    _it: Optional[Iterator[ConsumerRecord]]
-
-    def __init__(self) -> None:
-        # note: this is a regular dict, but ordered on Python 3.6
-        # we use this alias to signify it must be ordered.
-        self._buffers = OrderedDict()
-        # getmany calls next(_TopicBuffer), and does not call iter(),
-        # so the first call to next caches an iterator.
-        self._it = None
-
-    def add(self, tp: TP, buffer: List[ConsumerRecord]) -> None:
-        assert tp not in self._buffers
-        self._buffers[tp] = iter(buffer)
-
-    def __iter__(self) -> Iterator[Tuple[TP, ConsumerRecord]]:
-        buffers = self._buffers
-        buffers_items = buffers.items
-        buffers_remove = buffers.pop
-        sentinel = object()
-        to_remove: Set[TP] = set()
-        mark_as_to_remove = to_remove.add
-        while buffers:
-            for tp in to_remove:
-                buffers_remove(tp, None)
-            for tp, buffer in buffers_items():
-                item = next(buffer, sentinel)
-                if item is sentinel:
-                    mark_as_to_remove(tp)
-                    continue
-                yield tp, item
-
-    def __next__(self) -> Tuple[TP, ConsumerRecord]:
-        # Note: this method is not in normal iteration
-        # as __iter__ returns generator.
-        it = self._it
-        if it is None:
-            it = self._it = iter(self)
-        return it.__next__()
-
-
 class ConsumerRebalanceListener(aiokafka.abc.ConsumerRebalanceListener):
     # kafka's ridiculous class based callback interface makes this hacky.
 
-    def __init__(self, thread: 'ConsumerThread') -> None:
-        self._thread: 'ConsumerThread' = thread
+    def __init__(self, thread: ConsumerThread) -> None:
+        self._thread: ConsumerThread = thread
 
     async def on_partitions_revoked(
             self, revoked: Iterable[_TopicPartition]) -> None:
-        await self._thread.on_partitions_revoked(revoked)
+        await self._thread.on_partitions_revoked(ensure_TPset(revoked))
 
     async def on_partitions_assigned(
             self, assigned: Iterable[_TopicPartition]) -> None:
-        await self._thread.on_partitions_assigned(assigned)
+        await self._thread.on_partitions_assigned(ensure_TPset(assigned))
 
 
-class Consumer(base.Consumer):
+class Consumer(ThreadDelegateConsumer):
     """Kafka consumer using :pypi:`aiokafka`."""
     logger = logger
 
     RebalanceListener: ClassVar[Type[ConsumerRebalanceListener]]
     RebalanceListener = ConsumerRebalanceListener
 
-    _thread: 'ConsumerThread'
-    _active_partitions: Optional[Set[_TopicPartition]]
-    _paused_partitions: Set[_TopicPartition]
-    fetch_timeout: float = 10.0
-
     consumer_stopped_errors: ClassVar[Tuple[Type[BaseException], ...]] = (
         ConsumerStoppedError,
     )
 
-    flow_active: bool = True
-    can_resume_flow: Event
+    def _new_consumer_thread(self) -> ConsumerThread:
+        return AIOKafkaConsumerThread(self, loop=self.loop, beacon=self.beacon)
 
-    #: Main thread method queue.
-    #: The consumer is running in a separate thread, and so we send
-    #: requests to it via a queue.
-    #: Sometimes the thread needs to call code owned by the main thread,
-    #: such as App.on_partitions_revoked, and in that case the thread
-    #: uses this method queue.
-    _method_queue: MethodQueue
-
-    def on_init(self) -> None:
-        self._active_partitions = None
-        self._paused_partitions = set()
-        self.can_resume_flow = Event()
-        self._method_queue = MethodQueue(loop=self.loop, beacon=self.beacon)
-        self._thread = ConsumerThread(
-            self,
-            loop=self.loop,
-            beacon=self.beacon,
+    async def create_topic(self,
+                           topic: str,
+                           partitions: int,
+                           replication: int,
+                           *,
+                           config: Mapping[str, Any] = None,
+                           timeout: Seconds = 30.0,
+                           retention: Seconds = None,
+                           compacting: bool = None,
+                           deleting: bool = None,
+                           ensure_created: bool = False) -> None:
+        await self._thread.create_topic(
+            topic,
+            partitions,
+            replication,
+            config=config,
+            timeout=int(want_seconds(timeout) * 1000.0),
+            retention=int(want_seconds(retention) * 1000.0),
+            compacting=compacting,
+            deleting=deleting,
+            ensure_created=ensure_created,
         )
 
-    async def on_restart(self) -> None:
-        self.on_init()
+    def _new_topicpartition(self, topic: str, partition: int) -> TP:
+        return cast(TP, _TopicPartition(topic, partition))
 
-    def _get_active_partitions(self) -> Set[_TopicPartition]:
-        tps = self._active_partitions
-        if tps is None:
-            # need aiokafka._TopicPartition, not faust.TP
-            return self._set_active_tps(self.assignment())
-        return tps
+    def _to_message(self, tp: TP, record: Any) -> ConsumerMessage:
+        timestamp: Optional[int] = record.timestamp
+        timestamp_s: float = cast(float, None)
+        if timestamp is not None:
+            timestamp_s = timestamp / 1000.0
+        return ConsumerMessage(
+            record.topic,
+            record.partition,
+            record.offset,
+            timestamp_s,
+            record.timestamp_type,
+            record.key,
+            record.value,
+            record.checksum,
+            record.serialized_key_size,
+            record.serialized_value_size,
+            tp,
+        )
 
-    def _set_active_tps(self,
-                        tps: Set[_TopicPartition]) -> Set[_TopicPartition]:
-        tps = self._active_partitions = set(tps)  # copy!
-        tps.difference_update(self._paused_partitions)
-        return tps
+    async def on_stop(self) -> None:
+        await super().on_stop()
+        transport = cast(Transport, self.transport)
+        transport._topic_waiters.clear()
+
+
+class AIOKafkaConsumerThread(ConsumerThread):
+    _consumer: Optional[aiokafka.AIOKafkaConsumer] = None
+
+    def on_init(self) -> None:
+        self._rebalance_listener = self.consumer.RebalanceListener(self)
+
+    async def on_start(self) -> None:
+        self._consumer = self._create_consumer(loop=self.thread_loop)
+        await self._consumer.start()
 
     def _create_consumer(
             self,
@@ -250,370 +209,10 @@ class Consumer(base.Consumer):
             ssl_context=conf.ssl_context,
         )
 
-    async def create_topic(self,
-                           topic: str,
-                           partitions: int,
-                           replication: int,
-                           *,
-                           config: Mapping[str, Any] = None,
-                           timeout: Seconds = 30.0,
-                           retention: Seconds = None,
-                           compacting: bool = None,
-                           deleting: bool = None,
-                           ensure_created: bool = False) -> None:
-        await self._thread.create_topic(
-            topic,
-            partitions,
-            replication,
-            config=config,
-            timeout=int(want_seconds(timeout) * 1000.0),
-            retention=int(want_seconds(retention) * 1000.0),
-            compacting=compacting,
-            deleting=deleting,
-            ensure_created=ensure_created,
-        )
-
-    async def on_start(self) -> None:
-        await self.add_runtime_dependency(self._method_queue)
-        await self.add_runtime_dependency(self._thread)
-
-    async def threadsafe_partitions_revoked(
-            self,
-            receiver_loop: asyncio.AbstractEventLoop,
-            revoked: Set[TP]) -> None:
-        promise = await self._method_queue.call(
-            receiver_loop.create_future(),
-            self.on_partitions_revoked,
-            revoked,
-        )
-        # wait for main-thread to finish processing request
-        await promise
-
-    async def threadsafe_partitions_assigned(
-            self,
-            receiver_loop: asyncio.AbstractEventLoop,
-            assigned: Set[TP]) -> None:
-        promise = await self._method_queue.call(
-            receiver_loop.create_future(),
-            self.on_partitions_assigned,
-            assigned,
-        )
-        # wait for main-thread to finish processing request
-        await promise
-
-    async def subscribe(self, topics: Iterable[str]) -> None:
-        await self._thread.subscribe(topics=topics)
-
-    async def getmany(self,
-                      timeout: float) -> AsyncIterator[Tuple[TP, Message]]:
-        if not self.flow_active:
-            await self.wait(self.can_resume_flow)
-        # Implementation for the Fetcher service.
-        active_partitions = self._get_active_partitions()
-        _next = next
-
-        records: RecordMap = {}
-        if active_partitions:
-            # Fetch records only if active partitions to avoid the risk of
-            # fetching all partitions in the beginning when none of the
-            # partitions is paused/resumed.
-            records = await self._thread.getmany(
-                active_partitions,
-                timeout=timeout,
-            )
-        else:
-            # We should still release to the event loop
-            await self.sleep(1)
-            if self.should_stop:
-                return
-        create_message = ConsumerMessage  # localize
-
-        # records' contain mapping from TP to list of messages.
-        # if there are two agents, consuming from topics t1 and t2,
-        # normal order of iteration would be to process each
-        # tp in the dict:
-        #    for tp. messages in records.items():
-        #        for message in messages:
-        #           yield tp, message
-        #
-        # The problem with this, is if we have prefetched 16k records
-        # for one partition, the other partitions won't even start processing
-        # before those 16k records are completed.
-        #
-        # So we try round-robin between the tps instead:
-        #
-        #    iterators: Dict[TP, Iterator] = {
-        #        tp: iter(messages)
-        #        for tp, messages in records.items()
-        #    }
-        #    while iterators:
-        #        for tp, messages in iterators.items():
-        #            yield tp, next(messages)
-        #            # remove from iterators if empty.
-        #
-        # The problem with this implementation is that
-        # the records mapping is ordered by TP, so records.keys()
-        # will look like this:
-        #
-        #  TP(topic='bar', partition=0)
-        #  TP(topic='bar', partition=1)
-        #  TP(topic='bar', partition=2)
-        #  TP(topic='bar', partition=3)
-        #  TP(topic='foo', partition=0)
-        #  TP(topic='foo', partition=1)
-        #  TP(topic='foo', partition=2)
-        #  TP(topic='foo', partition=3)
-        #
-        # If there are 100 partitions for each topic,
-        # it will process 100 items in the first topic, then 100 items
-        # in the other topic, but even worse if partition counts
-        # vary greatly, t1 has 1000 partitions and t2
-        # has 1 partition, then t2 will end up being starved most of the time.
-        #
-        # We solve this by going round-robin through each topic.
-        topic_index = self._records_to_topic_index(records, active_partitions)
-        to_remove: Set[str] = set()
-        sentinel = object()
-        while topic_index:
-            if not self.flow_active:
-                break
-            for topic in to_remove:
-                topic_index.pop(topic, None)
-            for topic, messages in topic_index.items():
-                if not self.flow_active:
-                    break
-                item = _next(messages, sentinel)
-                if item is sentinel:
-                    # this topic is now empty,
-                    # but we cannot remove from dict while iterating over it,
-                    # so move that to the outer loop.
-                    to_remove.add(topic)
-                    continue
-                tp, record = item  # type: ignore
-                if tp in active_partitions:
-                    highwater_mark = self._thread.highwater(tp)
-                    self.app.monitor.track_tp_end_offset(tp, highwater_mark)
-                    # convert timestamp to seconds from int milliseconds.
-                    timestamp: Optional[int] = record.timestamp
-                    timestamp_s: float = cast(float, None)
-                    if timestamp is not None:
-                        timestamp_s = timestamp / 1000.0
-                    yield tp, create_message(
-                        record.topic,
-                        record.partition,
-                        record.offset,
-                        timestamp_s,
-                        record.timestamp_type,
-                        record.key,
-                        record.value,
-                        record.checksum,
-                        record.serialized_key_size,
-                        record.serialized_value_size,
-                        tp,
-                    )
-
-    def _records_to_topic_index(
-            self,
-            records: RecordMap,
-            active_partitions: Set[_TopicPartition]) -> TopicIndexMap:
-        topic_index: TopicIndexMap = {}
-        for tp, messages in records.items():
-            try:
-                entry = topic_index[tp.topic]
-            except KeyError:
-                entry = topic_index[tp.topic] = _TopicBuffer()
-            entry.add(tp, messages)
-        return topic_index
-
-    def _new_topicpartition(self, topic: str, partition: int) -> TP:
-        return cast(TP, _TopicPartition(topic, partition))
-
-    def _new_offsetandmetadata(self, offset: int, meta: str) -> Any:
-        return OffsetAndMetadata(offset, meta)
-
-    async def on_stop(self) -> None:
-        await super().on_stop()  # wait_empty
-        await self.commit()
-        transport = cast(Transport, self.transport)
-        transport._topic_waiters.clear()
-
-    async def perform_seek(self) -> None:
-        read_offset = self._read_offset
-        _committed_offsets = await self._thread.seek_to_committed()
-        committed_offsets = {
-            _ensure_TP(tp): offset
-            for tp, offset in _committed_offsets.items()
-            if offset is not None
-        }
-        read_offset.update({
-            tp: offset if offset else None
-            for tp, offset in committed_offsets.items()
-        })
-        self._committed_offset.update(committed_offsets)
-
-    async def _commit(self, offsets: Mapping[TP, Tuple[int, str]]) -> bool:
-        table = terminal.logtable(
-            [(str(tp), str(offset), meta)
-             for tp, (offset, meta) in offsets.items()],
-            title='Commit Offsets',
-            headers=['TP', 'Offset', 'Metadata'],
-        )
-        self.log.dev('COMMITTING OFFSETS:\n%s', table)
-        try:
-            assignment = self.assignment()
-            commitable: Dict[TP, OffsetAndMetadata] = {}
-            revoked: Dict[TP, OffsetAndMetadata] = {}
-            commitable_offsets: Dict[TP, int] = {}
-            for tp, (offset, meta) in offsets.items():
-                offset_and_metadata = self._new_offsetandmetadata(offset, meta)
-                if tp in assignment:
-                    commitable_offsets[tp] = offset
-                    commitable[tp] = offset_and_metadata
-                else:
-                    revoked[tp] = offset_and_metadata
-            if revoked:
-                self.log.info(
-                    'Discarded commit for revoked partitions that '
-                    'will be eventually processed again: %r',
-                    revoked,
-                )
-            if not commitable:
-                return False
-            with flight_recorder(self.log, timeout=300.0) as on_timeout:
-                on_timeout.info('+aiokafka_consumer.commit()')
-                await self._thread.commit(commitable)
-                on_timeout.info('-aiokafka._consumer.commit()')
-            self._committed_offset.update(commitable_offsets)
-            self.app.monitor.on_tp_commit(commitable_offsets)
-            self._last_batch = None
-            return True
-        except CommitFailedError as exc:
-            if 'already rebalanced' in str(exc):
-                return False
-            self.log.exception(f'Committing raised exception: %r', exc)
-            await self.crash(exc)
-            return False
-        except IllegalStateError as exc:
-            self.log.exception(f'Got exception: {exc}\n'
-                               f'Current assignment: {self.assignment()}')
-            await self.crash(exc)
-            return False
-
-    def stop_flow(self) -> None:
-        self.flow_active = False
-        self.can_resume_flow.clear()
-
-    def resume_flow(self) -> None:
-        self.flow_active = True
-        self.can_resume_flow.set()
-
-    def pause_partitions(self, tps: Iterable[TP]) -> None:
-        tpset = set(tps)
-        self._get_active_partitions().difference_update(tpset)
-        self._paused_partitions.update(tpset)
-
-    def resume_partitions(self, tps: Iterable[TP]) -> None:
-        tpset = set(tps)
-        self._get_active_partitions().update(tps)
-        self._paused_partitions.difference_update(tpset)
-
-    async def position(self, tp: TP) -> Optional[int]:
-        return await self._thread.position(tp)
-
-    async def seek_wait(self, partitions: Mapping[TP, int]) -> None:
-        return await self._thread.seek_wait(partitions)
-
-    async def _seek_to_beginning(self, *partitions: TP) -> None:
-        self.log.dev('SEEK TO BEGINNING: %r', partitions)
-        self._read_offset.update((_ensure_TP(tp), None) for tp in partitions)
-        await self._thread.seek_to_beginning(*(
-            self._new_topicpartition(tp.topic, tp.partition)
-            for tp in partitions
-        ))
-
-    async def seek(self, partition: TP, offset: int) -> None:
-        self.log.dev('SEEK %r -> %r', partition, offset)
-        # reset livelock detection
-        self._last_batch = None
-        # set new read offset so we will reread messages
-        self._read_offset[_ensure_TP(partition)] = offset if offset else None
-        self._thread.seek(partition, offset)
-
-    def assignment(self) -> Set[TP]:
-        return self._thread.assignment()
-
-    def highwater(self, tp: TP) -> int:
-        return self._thread.highwater(tp)
-
-    def topic_partitions(self, topic: str) -> Optional[int]:
-        return self._thread.topic_partitions(topic)
-
-    async def earliest_offsets(self,
-                               *partitions: TP) -> MutableMapping[TP, int]:
-        return await self._thread.earliest_offsets(*partitions)
-
-    async def highwaters(self, *partitions: TP) -> MutableMapping[TP, int]:
-        return await self._thread.highwaters(*partitions)
-
-    def close(self) -> None:
-        self._thread.close()
-
-
-class ConsumerThread(QueueServiceThread):
-    app: AppT
-    consumer: Consumer
-    _consumer: Optional[aiokafka.AIOKafkaConsumer] = None
-
-    def __init__(self, consumer: Consumer, **kwargs: Any) -> None:
-        self.consumer = consumer
-        transport = self.consumer.transport
-        self.app = transport.app
-        self._rebalance_listener = consumer.RebalanceListener(self)
-        super().__init__(**kwargs)
-
-    async def on_start(self) -> None:
-        self._consumer = self.consumer._create_consumer(loop=self.thread_loop)
-        await self._consumer.start()
-
     def close(self) -> None:
         if self._consumer is not None:
             self._consumer.set_close()
             self._consumer._coordinator.set_close()
-
-    async def on_partitions_revoked(
-            self, revoked: Iterable[_TopicPartition]) -> None:
-        self.consumer.app.rebalancing = True  # set as early as possible
-        # see comment in on_partitions_assigned
-        consumer = self.consumer
-        _revoked = cast(Set[TP], set(revoked))
-        # remove revoked partitions from active + paused tps.
-        if consumer._active_partitions is not None:
-            consumer._active_partitions.difference_update(_revoked)
-        consumer._paused_partitions.difference_update(_revoked)
-        # start callback chain of assigned callbacks.
-        await consumer.threadsafe_partitions_revoked(
-            self.thread_loop, _revoked)
-
-    async def on_partitions_assigned(
-            self, assigned: Iterable[_TopicPartition]) -> None:
-        # have to cast to Consumer since ConsumerT interface does not
-        # have this attribute (mypy currently thinks a Callable instance
-        # variable is an instance method).  Furthermore we have to cast
-        # the Kafka TopicPartition namedtuples to our description,
-        # that way they are typed and decoupled from the actual client
-        # implementation.
-        consumer = self.consumer
-        _assigned = set(assigned)
-        # remove recently revoked tps from set of paused tps.
-        consumer._paused_partitions.intersection_update(_assigned)
-        # cache set of assigned partitions
-        cast(Set[TP], consumer._set_active_tps(_assigned))
-        # start callback chain of assigned callbacks.
-        #   need to copy set at this point, since we cannot have
-        #   the callbacks mutate our active list.
-        consumer._last_batch = None
-        await consumer.threadsafe_partitions_assigned(
-            self.thread_loop, _assigned)
 
     async def subscribe(self, topics: Iterable[str]) -> None:
         # XXX pattern does not work :/
@@ -627,9 +226,28 @@ class ConsumerThread(QueueServiceThread):
         return await self.call_thread(
             self._ensure_consumer().seek_to_committed)
 
-    async def commit(self, tps: Any) -> Any:
-        return await self.call_thread(
-            self._ensure_consumer().commit, tps)
+    async def commit(self, offsets: Mapping[TP, int]) -> bool:
+        return await self.call_thread(self._commit, offsets)
+
+    async def _commit(self, offsets: Mapping[TP, int]) -> bool:
+        consumer = self._ensure_consumer()
+        try:
+            consumer.commit({
+                tp: OffsetAndMetadata(offset, '')
+                for tp, offset in offsets.items()
+            })
+        except CommitFailedError as exc:
+            if 'already rebalanced' in str(exc):
+                return False
+            self.log.exception(f'Committing raised exception: %r', exc)
+            await self.crash(exc)
+            return False
+        except IllegalStateError as exc:
+            self.log.exception(f'Got exception: {exc}\n'
+                               f'Current assignment: {self.assignment()}')
+            await self.crash(exc)
+            return False
+        return True
 
     async def position(self, tp: TP) -> Optional[int]:
         return await self.call_thread(
@@ -657,7 +275,7 @@ class ConsumerThread(QueueServiceThread):
         self._ensure_consumer().seek(partition, offset)
 
     def assignment(self) -> Set[TP]:
-        return cast(Set[TP], self._ensure_consumer().assignment())
+        return ensure_TPset(self._ensure_consumer().assignment())
 
     def highwater(self, tp: TP) -> int:
         return self._ensure_consumer().highwater(tp)
@@ -682,7 +300,7 @@ class ConsumerThread(QueueServiceThread):
         return self._consumer
 
     async def getmany(self,
-                      active_partitions: Set[_TopicPartition],
+                      active_partitions: Set[TP],
                       timeout: float) -> RecordMap:
         # Implementation for the Fetcher service.
         _consumer = self._ensure_consumer()
